ElasticSearchTarget - Added MaxRecursionLimit to protect against stack overflow from cyclic references (code layout)
